piRho-backend/requirements.txt
# piRho SaaS Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.3.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=41.0.0

# Monitoring
sentry-sdk[fastapi]>=1.39.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Backtesting
pandas>=2.1.0
numpy>=1.26.0

